Extract HiveMinioDataLake class
